Update Main.jsx
added new modules and update the previous one. ( update the handleCheckboxChange, and now in this module whenever i click or unclick on any item it'll automatically update the count, also added the same thing in the delete note module. And at last added the functionality of clear complete to the project.